change name of model scripts and update makefile, update usage scripts in readme
