+++ docs/05_generate_report.Rmd
@@ -60,7 +60,7 @@ knitr::include_graphics("../results/class_proportion_bl.png")
 
 ## Analysis
 
-The multinomial Naive Bayes (`MultinomialNB` (@sklearn) ) algorithm was used to build a classification model to predict whether a license plate was 'accepted' or 'rejected' (found in the outcome column of the data set). We chose to fit a `MultinomialNB` model because the strength of predictors is easily interpretable, and the model performs efficiently on high-dimension datasets. Based on the exploratory data analysis it was deemed important to optimize the range of n-grams used in `CountVectorizer` (@sklearn) during feature engineering. We hard-coded the `analyzer` of `CountVectorizer` to `char` in order to create features without a word-boundary restirction. Then, using a model pipeline to optimize `ngram_range` using `GridSearchCV`, (@sklearn) the optimum model was shown to have n-grams in the range of `(2,3)`. On inspection of cross-walidated training error compared to the validation error for different `ngram_range` values, we can see that there are better optimal n-gram ranges than `(2,3)` when we consider the number of features. As shown below, `(2,2)` is likely a more optimal `ngram_range` given that it achieves similar accuracy scores, is less overfit, and includes significantly fewer features (over 6,000 fewer) than the `(2,3)` range. 
+The multinomial Naive Bayes (`MultinomialNB` (@sklearn) ) algorithm was used to build a classification model to predict whether a license plate was 'accepted' or 'rejected' (found in the outcome column of the data set). We chose to fit a `MultinomialNB` model because the strength of predictors is easily interpretable, and the model performs efficiently on high-dimension datasets. Based on the exploratory data analysis it was deemed important to optimize the range of n-grams used in `CountVectorizer` (@sklearn) during feature engineering. We hard-coded the `analyzer` of `CountVectorizer` to `char` in order to create features without a word-boundary restirction. Then, using a model pipeline to optimize `ngram_range` using `GridSearchCV`, (@sklearn) the optimum model was shown to have n-grams in the range of `(2,3)` (all n-grams of length 2 to length 3). On inspection of cross-walidated training error compared to the validation error for different `ngram_range` values, we can see that there are better optimal n-gram ranges than `(2,3)` when we consider the number of features. As shown below, `(2,2)` is likely a more optimal `ngram_range` given that it achieves similar accuracy scores, is less overfit, and includes significantly fewer features (over 6,000 fewer) than the `(2,3)` range. 
 
 ```{r ngram training and validation error, echo=FALSE, fig.align='center'}
 knitr::include_graphics("../results/train_val_error.png")
@@ -74,11 +74,11 @@ Validation evaluation metrics:
 knitr::include_graphics("../results/clf_val_report.png")
 ```
 
-Though the recall for the rejected class is fairly low, it's precision is decent. This is not very concerning, as in the context of license plate submissions, false negatives can be considered more harmful than fasle positives.  As an applicant, I would rather be rejected in the initial screening phase only if I was positive that my application would be rejected in the second screening phase. If the initial screening phase was overly conservative, ther would be no need for a secondary screening phase, and any application would be more likely to be rejected even if it were completely acceptable.
+Though the recall for the rejected class is fairly low, it's precision is decent. This is not very concerning, as in the context of license plate submissions, false negatives can be considered more harmful than false positives.  As an applicant, I would rather be rejected in the initial screening phase only if I was positive that my application would be rejected in the second screening phase. If the initial screening phase was overly conservative, ther would be no need for a secondary screening phase, and any application would be more likely to be rejected even if it were completely acceptable.
 
 # Results
 
-The optimized model shows simlar accuracy on the testing dataset. As seen below, our model achieves a testing accuracy score of 0.7397 overall. Like the validation metrics showed, the model predicts the `accepted` class better than it predicts the `rejected class` which is not ideal, but it also has a higher precision than recall, which, as discussed above, it preferable in this context.
+The optimized model shows similar accuracy on the testing dataset. As seen below, our model achieves a testing accuracy score of 0.7397 overall. Like the validation metrics showed, the model predicts the `accepted` class better than it predicts the `rejected class` which is not ideal, but it also has a higher precision than recall, which, as discussed above, it preferable in this context.
 
 Testing evaluation metrics
 ```{r eval metrics testing, echo=FALSE, fig.align='center'}
